Harden logging target resolution on 1.3
