+++ news/3418.bugfix
@@ -1,0 +1,1 @@
+Apply Hydra's target blocklist to Python logging configuration.
